// src/app/models/portfolio.ts
export class Portfolio {
    id: string;
    name: string;
    date: any;
    number: string;
    onlineid: string;
    passwordhint: string;
    memorableplace: string;
    memorablename: string;
    memorabledate: string;
    charge: number;
    initialcash: number;
    initialstocksbalance: number;
    initialtrustsbalance: number;
    taxable: string;
    stocksbalance: number;
    trustsbalance: number;
    balance: number;
    notes: string;
    forecast: boolean;
}